modification and debugging
