--- ravenous/public/robots.txt
@@ -1,2 +1,0 @@
-# https://www.robotstxt.org/robotstxt.html
-User-agent: *
