+++ ravenous/src/components/App/App.css
@@ -1,3 +1,5 @@
+/* Universal Styles */
+
 html {
   font-size: 18px;
 }
