+++ ravenous/src/components/Business/Business.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import './Business.css';
-import ReactDOM from 'react-dom';
-
 
 const business = {
     imageSrc: 'https://s3.amazonaws.com/codecademy-content/programs/react/ravenous/pizza.jpg',
