+++ ravenous/src/components/SearchBar/SearchBar.js
@@ -8,11 +8,11 @@ const sortByOptions = {
 };
 
 
-class SearchBar extends React.Components {
+class SearchBar extends React.Component {
    renderSortByOptions() {
        return Object.keys(sortByOptions).map(sortByOption => {
         let sortByOptionValue = sortByOptions[sortByOption];
-       return <li key={sortByOptionValue}>{sortByOption}</li>
+       return <li key={sortByOptionValue}>{sortByOption}</li>;
        });
    }
    render(){
